backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx
supabase==2.0.2
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
bcrypt